Put together the first stable version of the app
The UI was thought through and build up using React functional
components and Hooks.

The trickiest part was dealing with the click coords, which ended up
restricting the responsiveness of the app since the values were
measured in pixels, hence different window sizes may get in trouble
playing the game.

Conditional style rendering was made use of, so as to draw a square
on the click spot and push the character guess selection either to the
left or to the right, depending on which side of the screen was clicked.

The highscores are stored and retrieved using the Firestore database, a
feature offered by Firebase, a Google BAAS.